refactor: sockets/index.ts 내의 코드블럭 ClientIOTypes.ts로 이동
